developAnxo <- Master

+++ src/main/java/com/tsw/CompayRest/Controller/ExpenseController.java
@@ -1,20 +1,16 @@
 package com.tsw.CompayRest.Controller;
 
 import com.tsw.CompayRest.Dto.*;
-import com.tsw.CompayRest.Mapper.ExpenseMapperImpl;
-import com.tsw.CompayRest.Model.ExpenseModel;
 import com.tsw.CompayRest.Service.ExpenseService;
 import com.tsw.CompayRest.Service.ExpenseShareService;
 import com.tsw.CompayRest.Service.GroupService;
 import com.tsw.CompayRest.Service.UserService;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.*;
-import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/users/{userId}/groups/{groupId}/expenses")
@@ -23,14 +19,12 @@ public class ExpenseController {
     private final GroupService groupService;
     private final UserService userService;
     private final ExpenseShareService expenseShareService;
-    private final ExpenseMapperImpl expenseMapperImpl;
 
-    public ExpenseController(ExpenseService expenseService, GroupService groupService, UserService userService, ExpenseShareService expenseShareService, ExpenseMapperImpl expenseMapperImpl) {
+    public ExpenseController(ExpenseService expenseService, GroupService groupService, UserService userService, ExpenseShareService expenseShareService) {
         this.expenseService = expenseService;
         this.groupService = groupService;
         this.userService = userService;
         this.expenseShareService = expenseShareService;
-        this.expenseMapperImpl = expenseMapperImpl;
     }
 
     @GetMapping
@@ -68,7 +62,7 @@ public class ExpenseController {
         if(group.isPresent() && user.isPresent()) {
             ExpenseDto expenseDto = new ExpenseDto();
             expenseDto.setExpense_name(expense.getExpense_name());
-            expenseDto.setExpense_date(LocalDate.now());
+            expenseDto.setExpense_date(LocalDateTime.now());
             expenseDto.setAmount(expense.getAmount());
             expenseDto.setShare_method(expense.getShare_method());
             expenseDto.setOrigin_user(user.get());
@@ -119,13 +113,20 @@ public class ExpenseController {
     }
 
     @DeleteMapping("/{expenseId}")
-    public ResponseEntity<Void> deleteExpense(@PathVariable("expenseId") Long expenseId) {
-        boolean deletedShare = expenseShareService.deleteAll(expenseId);
-        boolean deletedExp = expenseService.deleteExpense(expenseId);
-        if (deletedExp && deletedShare) {
-            return ResponseEntity.noContent().build();
-        } else {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<Void> deleteExpense(@PathVariable("groupId") Long groupId, @PathVariable("expenseId") Long expenseId) {
+        Optional<ExpenseDto> toDelete = expenseService.getExpenseById(expenseId);
+
+        if(toDelete.isPresent()) {
+            boolean deletedShare = expenseShareService.deleteAll(expenseId);
+            boolean deletedExp = expenseService.deleteExpense(expenseId);
+
+            if(deletedShare && deletedExp) {
+                groupService.updateGroupAmount(groupId,-toDelete.get().getAmount());
+                return ResponseEntity.noContent().build();
+            }
+
         }
+
+        return ResponseEntity.notFound().build();
     }
 }
